// src/client/src/models/ExerciseSet.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WorkoutTracker
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 *
 * @export
 * @interface ExerciseSet
 */
export interface ExerciseSet {
  /**
   *
   * @type {number}
   * @memberof ExerciseSet
   */
  reps?: number;
  weight?: number;
}

/**
 * Check if a given object implements the ExerciseSet interface.
 */
export function instanceOfExerciseSet(value: object): boolean {
  let isInstance = true;

  return isInstance;
}

export function ExerciseSetFromJSON(json: any): ExerciseSet {
  return ExerciseSetFromJSONTyped(json, false);
}

export function ExerciseSetFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): ExerciseSet {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    reps: !exists(json, 'reps') ? undefined : json['reps'],
    weight: !exists(json, 'weight') ? undefined : json['weight'],
  };
}

export function ExerciseSetToJSON(value?: ExerciseSet | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    reps: value.reps,
    weight: value.weight,
  };
}